requirements.txt
# Core dependencies
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0
click>=8.0.0

# LLM APIs
openai>=1.0.0
requests>=2.28.0

# Database
redis>=4.5.0

# Data handling
pandas>=2.0.0
faker>=20.0.0  # For generating realistic sample data

# Utilities
python-dateutil>=2.8.0
colorama>=0.4.6  # For colored terminal output 